Fix failing prefix lookup test
Strangely, the reverse lookup for http://www.w3.org/2004/02/skos/core#
does not work. It returns only dc. This can be reproduced via the
browser as well, which redirects to a page of alternative results
instead of SKOS.

Try: http://prefix.cc/reverse?uri=http://xmlns.com/foaf/0.1/&format=json

A JSON result is returned as expected with expected items.

Then try:
http://prefix.cc/reverse?uri=http://www.w3.org/2004/02/skos/core%23&format=json

A JSON result is returned as expected, but it contains only the
(unexpected) dc prefix and corresponding namespace.

Go on to http://prefix.cc/ and enter
http://www.w3.org/2004/02/skos/core# in the search bar. The redirected
page shows dc as the canonical and lists skos as the last alternative
(as of date).

Somebody should report this but for now, fall back to using the DC pair
that's being returned, as the test.