Soft-lock dependency versions
Lock to the latest packages of the lowest-supported Python (3.8).
Explicit dependencies only, so as to not confuse people (which is what
`pip freeze` is often good at doing). Use a trick like below to
manually/ChatGPT inspect/replace existing requirements:

`grep -wiFf $requirements_file <(pip freeze)`

Remove version specifiers in the requirement files and upgrade
dependencies first, then compare the two lists. Pip dependencies like
`setuptools` and `wheel` can be and are omitted. We might want to move
to listing only parent packages, or to Poetry :)