Support CI build on Mac and Windows
Attain true cross-platform compatibility by making Github Actions CI run
for all three major operating systems, for all runtimes.

- Add multi-dimensional environments in CI config
- Add job concurrency group (or risk being killed)

**Caveat**: Tox appears to run twice on Windows, and sometimes on Mac.
This may be because of the system-wide availability of multiple
interpreters in some runner types. We either need to decouple Tox from
the build (take it out of Makefile) and use the Tox GH Actions plug-in,
or add some condition to the job to run Tox for only one
runtime/interpreter version.