New version 0.2.0, dropping support for Python <= 3.7
Python 3.7 and earlier reached EOL earlier this year, and many
tools/libraries have stopped supporting those runtime. Continuing to
support an old runtime will be near impossible due to dependency hell.
It is therefore only wise to follow suit.

- Update code, tests
- Update CI config
- Update dependencies
- Update metadata

We will aim to support and test against the oldest `security` release
and the latest `bugfix` release, less one (if it's too new and not a lot
of tools have adopted it).

The jump from Python 3.6 to 3.8 doesn't involve that many changes among
dependency APIs for this project -- there is just a
`collections.Mapping` namespace change and a Pandas
`DataFrame.replace()` field behaviour change.

Users are encouraged to upgrade even if there is no API or ABI change in
this project itself, as an EOL runtime is a security risk.